Conflicts:
	SpriteBuilder/SpriteBuilder.xcodeproj/project.pbxproj

+++ SpriteBuilder/SpriteBuilder Tests/PackageImporter_Tests.m
@@ -71,7 +71,7 @@
     XCTAssertNotNil(error);
 
     NSArray *errors = error.userInfo[@"errors"];
-    XCTAssertEqual(errors.count, 1);
+    XCTAssertEqual((int)errors.count, 1);
     XCTAssertEqual(error.code, SBImportingPackagesError);
     NSError *underlyingError = errors[0];
     XCTAssertEqual(underlyingError.code, SBPackageAlreayInProject);
